+++ app/src/main/java/com/codesoom/assignment/application/TaskService.java
@@ -1,9 +1,7 @@
 package com.codesoom.assignment.application;
 
-import com.codesoom.assignment.models.Task;
 import com.codesoom.assignment.models.TaskDto;
 
-import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.List;
 
@@ -20,18 +18,4 @@ public interface TaskService {
     TaskDto updateTitle(Long id, TaskDto dto);
 
     TaskDto deleteTask(Long id);
-
-    default TaskDto taskToDto(Task task) {
-        return TaskDto.from(task);
-    }
-
-    default Task dtoToTask(Long id, TaskDto dto) {
-        return Task.builder()
-                .id(id)
-                .title(dto.getTitle())
-                .regDate(LocalDateTime.now())
-                .modDate(LocalDateTime.now())
-                .deadLine(dto.getDeadLine())
-                .build();
-    }
 }
