+++ app/src/main/java/com/codesoom/assignment/utils/TaskIdGenerator.java
@@ -1,15 +1,6 @@
 package com.codesoom.assignment.utils;
 
-import org.springframework.stereotype.Component;
+public interface TaskIdGenerator {
 
-import java.util.concurrent.atomic.AtomicLong;
-
-@Component
-public class TaskIdGenerator {
-
-    private final AtomicLong atomicLong = new AtomicLong(1L);
-
-    public Long generateId() {
-        return atomicLong.getAndIncrement();
-    }
+    Long generateId();
 }
