+++ app/src/main/java/com/codesoom/assignment/utils/TaskIdGeneratorImpl.java
@@ -1,0 +1,15 @@
+package com.codesoom.assignment.utils;
+
+import org.springframework.stereotype.Component;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+@Component
+public class TaskIdGeneratorImpl implements TaskIdGenerator {
+
+    private final AtomicLong atomicLong = new AtomicLong(1L);
+
+    public Long generateId() {
+        return atomicLong.getAndIncrement();
+    }
+}
